Move helper functions to itineraryHelpers.js
